+++ examples/nextjs/pages/middleware-protected/_middleware.ts
@@ -1,0 +1,3 @@
+import { withMiddlewareAuthRequired } from '@supabase/supabase-auth-helpers/nextjs';
+
+export const middleware = withMiddlewareAuthRequired();